backend/requirements.txt
# Core Python Libraries
python-dotenv==1.0.0

# Web Framework
flask==2.3.2
flask-cors==4.0.0

# AI and Machine Learning Libraries
openai==1.3.7
langchain-groq==0.1.0
tiktoken==0.5.1

# Optional but recommended for JSON handling
simplejson==3.19.1

# HTTP and API support
requests==2.31.0

# Optional utilities
pydantic==2.5.2

# If you want to use additional data processing (optional)
pandas==2.0.1
numpy==1.24.3

# Typing support
typing-extensions==4.8.0

# Async support (optional)
asyncio==3.4.3
